requirements.txt
torch
torchvision
tensorboard
pynvml
tqdm
pytorch-ignite
